+++ src/uk/ac/nott/mrl/homework/server/PermitDevice.java
@@ -28,10 +28,10 @@ public class PermitDevice extends HttpServlet
 		conn.setRequestMethod("POST");
 
 		final String sinceString = request.getParameter("since");
-		double since = 0;
+		long since = 0;
 		try
 		{
-			since = Double.parseDouble(sinceString);
+			since = Long.parseLong(sinceString);
 		}
 		catch (final Exception e)
 		{
